SIMP-348 Fix for :ref: issue with rst2pdf
This is a workaround for the current issue with ref2pdf where a PDF is
not generated if there are :ref: tags in your code.

In this case, the :ref: tags are stripped from the RST files prior to
generating the PDF.

This should be corrected when the bug is fixed.

This also fixes the issue with RTD failing.

SIMP-348 #close #comment 5.1.X Fix