(SIMP-460) Mainline SIMP documentation
This commit supports all of the SIMP versions under a single master
documentation branch.

We've been 90% there for quite some time but this finalizes the move.

Once this is committed, all simp-core branches should start using the
'master' branch of simp-doc moving forward.

The Changelog.rst for a given version should be at the top level of
simp-core.

SIMP-460 #close