Update patch set 1
Patch Set 1:

So removing the Changelog.rst from the root of the documentation causes the RPM build to fail.  The spec file is expecting it in the root.  Yous say it should be at the top level for give distro but it only lives in the common directory right now. C

Patch-set: 1
Reviewer: Ralph Wright <1007363@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0