(SIMP-1419) Restored and updated the SecCONOP
The new security mapping did not have the depth of coverage that the
old SecCONOP did, therefore it has been restored to work alongside the
mapping and to ensure that adequate information is covered.

The document has been updated and as many policy references as possible
were actively cross-linked against the NIST 800-53.

SIMP-1419 #close