(SIMP-1218) Fix issue with losing RTD information
The previous push to SIMP-1218 caused issues with with information loss
when publishing to ReadTheDocs. This ensures that as much information
as possible is pulled from GitHub when no other sources are available.

Much of this code should probably be consolidated into functions at
some point.

SIMP-1218 #close Fix information loss in RTD