(SIMP-3811) SIMP_FAST_DOCS=true flag added
Added support for a ``SIMP_FAST_DOCS`` environment variable that, when
set to ``true``, will exclude certain long-running sections of the
documentation from being included.

In general, this has shown a 66% speed up in the RPM build process but
should only be used for testing the RPM build to identify catastrophic
failures.