(SIMP-4546) Document unavoidable `ca` warning
Puppet 5.5.6+ marks the `[master]` section's `ca` setting as deprecated,
causing every `puppet` run to emit an eye-catching warning when the
configuration is set in `puppet.conf`.  **simp/pupmod** has been updated
to accomodate this, but in some circumstances it is unavoidable.

This patch documents the issue and provides a (heavy-handed) workaround.
It also draws attention to the issue in the 6.3.0 upgrade guide.

SIMP-5456 #close