(SIMP-6328) Added official hardware requirement for RAM plus swap
With less than 3.4 GB of RAM+swap the puppet database and puppet server
may not start. /etc/sysconfig/puppetdb requires 1516m and
/etc/sysconfig/puppetserver requires 1896m. This comes out to just over
3.3 GB.

If -Xms is lowered, it is possible to have a machine with less than 3.4
GB of RAM+swap, but performance will be impacted, and at some point
there will be too little allocated memory for the puppet database/server
to run.